@draveness 2019-11-03
/retest


Regards,
Draven
On Nov 3, 2019, 12:01 PM +0800, Kubernetes Prow Robot <notifications@github.com>, wrote:
> @draveness: The following tests failed, say /retest to rerun them all:
> Test name
> Commit
> Details
> Rerun command
> pull-kubernetes-e2e-gce-100-performance
> 2d2ff2a
> link
> /test pull-kubernetes-e2e-gce-100-performance
> pull-kubernetes-kubemark-e2e-gce-big
> 2d2ff2a
> link
> /test pull-kubernetes-kubemark-e2e-gce-big
> pull-kubernetes-verify
> 2d2ff2a
> link
> /test pull-kubernetes-verify
> pull-kubernetes-e2e-gce
> 786a80b
> link
> /test pull-kubernetes-e2e-gce
> Full PR test history. Your PR dashboard. Please help us cut down on flakes by linking to an open issue when you hit one in your PR.
> Instructions for interacting with me using PR comments are available here. If you have questions or suggestions related to my behavior, please file an issue against the kubernetes/test-infra repository. I understand the commands that are listed here.
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or unsubscribe.