@ahg-g 2019-11-04
I benchmarked the performance of this path, it is 16% slower:

Without:

```
BenchmarkScheduling/5000Nodes/1000Pods-12           1000           6112369 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           8190366 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           7201645 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000          13728735 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000          12725358 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000          11191880 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000          12539258 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000          11461610 ns/op
```

With this patch:

```
BenchmarkScheduling/5000Nodes/1000Pods-12           1000           7185840 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           9293825 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           8141612 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000          14786628 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000          14599050 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000          12326200 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000          13832831 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000          12638641 ns/op
```

I will try to do some profiling to figure out how to get it closer to the original performance.