@ahg-g 2019-11-05
I set https://github.com/kubernetes/kubernetes/pull/84824, it should address the performance issue of this PR.