@draveness 2019-11-06
> lets merge this so that you can start woking on the one that creates the corresponding Score plugin, but please address the two comments I left in the Score PR.

Thanks, will do