@ahg-g 2019-11-07
> There is a problem with the typecheck CI #84887

Go version update, Jordan sent an email saying PRs will fail the test until the upgrade PR is merged.